chore: #101 make files for node-move getters and utils to separate files
